+++ server/iocContainer/index.ts
@@ -1499,6 +1499,8 @@ export default async function getBottle(
       async (params) =>
         container.NcmecService.getUserHasExistingNcmecReport(params),
       container.ManualReviewContentResolver,
+      container.Meter,
+      process.env.MANUAL_REVIEW_METRICS_ORG_ID,
     );
   });